requirements.txt
# GUI
PyQt5==5.15.10

# Scraping
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
newspaper3k==0.2.8
lxml==5.1.0
lxml_html_clean

# Data & Export
pandas==2.2.0
openpyxl==3.1.2

# Utility
requests==2.31.0